Refactor points logic into usePoints
