+++ src/features/points/usePoints.tsx
@@ -1,10 +1,34 @@
 import assert from 'assert';
 
+import { useQuery } from 'react-query';
+
 import { client } from '../../lib/gql/client';
 
-import { getAvailablePoints } from './getAvailablePoints';
+import { POINTS_PER_GIVE, getAvailablePoints } from './getAvailablePoints';
 
-export const getMyAvailablePoints = async () => {
+export const POINTS_QUERY_KEY = 'points_query_key';
+
+export const usePoints = () => {
+  const { data } = useQuery(
+    [POINTS_QUERY_KEY],
+    async () => {
+      return await getMyAvailablePoints();
+    },
+    {
+      onError: error => {
+        console.error(error);
+      },
+    }
+  );
+
+  const points = data || 0;
+  const give = Math.floor(points / POINTS_PER_GIVE);
+  const canGive = give >= 1;
+
+  return { points, give, canGive };
+};
+
+const getMyAvailablePoints = async () => {
   const { profiles_private } = await client.query(
     {
       profiles_private: [
